Refactor lineup hooks
